+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AbstractIntegrationTest.java
@@ -112,8 +112,8 @@ abstract class AbstractIntegrationTest {
 			break;
 		case UTIL:
 			if (this.configuration.isSeparateUtilityClasses()) {
-				fullClassname = this.configuration.getPackageName() + "."
-						+ GenerateCodeDocumentParser.UTIL_PACKAGE_NAME + "." + simpleClassname;
+				fullClassname = this.configuration.getPackageName() + "." + GenerateCodeDocumentParser.UTIL_PACKAGE_NAME
+						+ "." + simpleClassname;
 			} else {
 				fullClassname = this.configuration.getPackageName() + "." + simpleClassname;
 			}