+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AllGraphQLCasesServer_util_allFalse_Test.java
@@ -1,5 +1,6 @@
 package com.graphql_java_generator.plugin.compilation_tests;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -7,6 +8,7 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.lang.reflect.Field;
+import java.util.Arrays;
 import java.util.List;
 
 import javax.persistence.Entity;
@@ -17,15 +19,16 @@ import javax.persistence.Transient;
 import org.dataloader.BatchLoaderEnvironment;
 import org.dataloader.DataLoader;
 import org.junit.jupiter.api.BeforeEach;
+import org.springframework.graphql.data.method.annotation.BatchMapping;
 import org.springframework.graphql.execution.BatchLoaderRegistry;
 
-import graphql.mavenplugin_notscannedbyspring.AllGraphQLCases_Server_SpringConfiguration_separateUtilityClasses;
+import graphql.mavenplugin_notscannedbyspring.AllGraphQLCases_Server_SpringConfiguration_util_allFalse;
 import graphql.schema.DataFetchingEnvironment;
 
-class AllGraphQLCasesServer_separateUtilityClasses_Test extends AbstractIntegrationTest {
+class AllGraphQLCasesServer_util_allFalse_Test extends AbstractIntegrationTest {
 
-	public AllGraphQLCasesServer_separateUtilityClasses_Test() {
-		super(AllGraphQLCases_Server_SpringConfiguration_separateUtilityClasses.class);
+	public AllGraphQLCasesServer_util_allFalse_Test() {
+		super(AllGraphQLCases_Server_SpringConfiguration_util_allFalse.class);
 	}
 
 	@BeforeEach
@@ -35,6 +38,15 @@ class AllGraphQLCasesServer_separateUtilityClasses_Test extends AbstractIntegrat
 
 	@Override
 	protected void doAdditionalChecks() throws Exception {
+		Class<?> queryControllerClass = loadGeneratedClass("MyQueryTypeController", FileType.UTIL);
+		assertNotNull(queryControllerClass);
+		//
+		Class<?> mutationControllerClass = loadGeneratedClass("AnotherMutationTypeController", FileType.UTIL);
+		assertNotNull(mutationControllerClass);
+		//
+		Class<?> subscriptionControllerClass = loadGeneratedClass("TheSubscriptionTypeController", FileType.UTIL);
+		assertNotNull(subscriptionControllerClass);
+		//
 		Class<?> allFieldCasesClass = loadGeneratedClass("AllFieldCases", FileType.POJO);
 		assertNotNull(allFieldCasesClass);
 		//
@@ -66,6 +78,31 @@ class AllGraphQLCasesServer_separateUtilityClasses_Test extends AbstractIntegrat
 						List.class, //
 						BatchLoaderEnvironment.class),
 				"There should be no batchLoader(List,BatchLoaderEnvironment) method");
+
+		//////////////////////////////////////////////////////////////////////////////////////
+		// generateBatchMappingDataFetchers
+		assertFalse(this.configuration.isGenerateBatchMappingDataFetchers(),
+				"generateBatchMappingDataFetchers should be false in this test");
+		//
+		// query, mutation and subscription should have no method with @BatchMapping annotation
+		assertEquals(0,
+				Arrays.stream(queryControllerClass.getDeclaredMethods())
+						.filter(m -> m.getAnnotation(BatchMapping.class) != null).count(),
+				"query, mutation and subscription should have no method with @BatchMapping annotation");
+		assertEquals(0,
+				Arrays.stream(mutationControllerClass.getDeclaredMethods())
+						.filter(m -> m.getAnnotation(BatchMapping.class) != null).count(),
+				"query, mutation and subscription should have no method with @BatchMapping annotation");
+		assertEquals(0,
+				Arrays.stream(subscriptionControllerClass.getDeclaredMethods())
+						.filter(m -> m.getAnnotation(BatchMapping.class) != null).count(),
+				"query, mutation and subscription should have no method with @BatchMapping annotation");
+		//
+		// No method in the controllers and DataFetcherDelegates should have the @BatchMapping annotation
+		assertEquals(0,
+				Arrays.stream(allFieldCasesControllerClass.getDeclaredMethods())
+						.filter(m -> m.getAnnotation(BatchMapping.class) != null).count(),
+				"type's controllers have no method with @BatchMapping annotation");
 
 		//////////////////////////////////////////////////////////////////////////////////////
 		// generateDataFetcherForEveryFieldsWithArguments