+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/generate_code/DocumentParser_Forum_Server_Test.java
@@ -105,7 +105,7 @@ class DocumentParser_Forum_Server_Test {
 		checkFieldAnnotation(topic.getFields().get(i++), "date",
 				"@GraphQLScalar( fieldName = \"date\", graphQLTypeSimpleName = \"Date\", javaClass = java.util.Date.class, listDepth = 0)");
 		checkFieldAnnotation(topic.getFields().get(i++), "author",
-				"@Transient\n\t@GraphQLNonScalar( fieldName = \"author\", graphQLTypeSimpleName = \"Member\", javaClass = org.graphql.mavenplugin.junittest.forum_server_springconfiguration.Member.class, listDepth = 0)");
+				"@Transient\n\t@GraphQLNonScalar( fieldName = \"author\", graphQLTypeSimpleName = \"Member\", javaClass = org.junittest.forum_server_springconfiguration.Member.class, listDepth = 0)");
 		checkFieldAnnotation(topic.getFields().get(i++), "publiclyAvailable",
 				"@GraphQLScalar( fieldName = \"publiclyAvailable\", graphQLTypeSimpleName = \"Boolean\", javaClass = java.lang.Boolean.class, listDepth = 0)");
 		checkFieldAnnotation(topic.getFields().get(i++), "nbPosts",
@@ -115,7 +115,7 @@ class DocumentParser_Forum_Server_Test {
 		checkFieldAnnotation(topic.getFields().get(i++), "content",
 				"@GraphQLScalar( fieldName = \"content\", graphQLTypeSimpleName = \"String\", javaClass = java.lang.String.class, listDepth = 0)");
 		checkFieldAnnotation(topic.getFields().get(i++), "posts",
-				"@Transient\n\t@GraphQLNonScalar( fieldName = \"posts\", graphQLTypeSimpleName = \"Post\", javaClass = org.graphql.mavenplugin.junittest.forum_server_springconfiguration.Post.class, listDepth = 1)");
+				"@Transient\n\t@GraphQLNonScalar( fieldName = \"posts\", graphQLTypeSimpleName = \"Post\", javaClass = org.junittest.forum_server_springconfiguration.Post.class, listDepth = 1)");
 	}
 
 	private void checkFieldAnnotation(Field field, String name, String annotation) {