+++ graphql-maven-plugin-logic/src/test/java/graphql/mavenplugin_notscannedbyspring/AllGraphQLCases_Server_SpringConfiguration_util_allFalse.java
@@ -28,7 +28,7 @@ import com.graphql_java_generator.plugin.test.helper.GraphQLConfigurationTestHel
 		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateRelaySchema.*"),
 		@Filter(type = FilterType.REGEX, pattern = ".*\\.GenerateGraphQLSchema.*"),
 		@Filter(type = FilterType.REGEX, pattern = "com.graphql_java_generator.client.graphqlrepository.*") })
-public class AllGraphQLCases_Server_SpringConfiguration_separateUtilityClasses2 extends AbstractSpringConfiguration {
+public class AllGraphQLCases_Server_SpringConfiguration_util_allFalse extends AbstractSpringConfiguration {
 
 	static List<CustomScalarDefinition> customScalars;
 	static {
@@ -63,24 +63,24 @@ public class AllGraphQLCases_Server_SpringConfiguration_separateUtilityClasses2 
 		configuration.customScalars = customScalars;
 
 		// Parameters that control the server generation
-		configuration.addRelayConnections = false; // same as in
-													// AllGraphQLCases_Server_SpringConfiguration_separateUtilityClasses
-		configuration.generateBatchLoaderEnvironment = true;// false testé dans separateUtilityClasses_Test
-		configuration.generateDataFetcherForEveryFieldsWithArguments = true;// false testé dans
-																			// separateUtilityClasses_Test
-		configuration.generateDataLoaderForLists = true;// false testé dans separateUtilityClasses_Test
-		configuration.generateJPAAnnotation = true;// false testé dans separateUtilityClasses_Test
-		configuration.generateUtilityClasses = true;
+		configuration.addRelayConnections = false;
+		configuration.generateBatchLoaderEnvironment = false; // true tesTed in AllGraphQLCasesServer_util_allTrue_Test
+		configuration.generateBatchMappingDataFetchers = false;// true is tested in
+		// AllGraphQLCasesServer_util_batchMapping_Test
+		configuration.generateDataFetcherForEveryFieldsWithArguments = false;// true tested in
+																				// AllGraphQLCasesServer_util_allTrue_Test
+		configuration.generateDataLoaderForLists = false;// true tesTed in AllGraphQLCasesServer_util_allTrue_Test
+		configuration.generateJPAAnnotation = false;// true tesTed in AllGraphQLCasesServer_util_allTrue_Test
+		configuration.generateUtilityClasses = true; // must be true for server implementation
 		configuration.ignoredSpringMappings = "";
 		configuration.javaTypeForIDType = GenerateServerCodeConfiguration.DEFAULT_JAVA_TYPE_FOR_ID_TYPE;
 		configuration.queryMutationExecutionProtocol = QueryMutationExecutionProtocol.http;
-		configuration.separateUtilityClasses = false;
+		configuration.separateUtilityClasses = true;
 
 		// Other parameters
 		configuration.mode = PluginMode.server;
 		configuration.schemaFilePattern = "allGraphQLCases*.graphqls";
 		configuration.schemaPersonalizationFile = new File(this.mavenTestHelper.getModulePathFile(),
 				"src/test/resources/schema_personalization/schema_personalization_for_code_generation.json");
-		configuration.separateUtilityClasses = true;
 	}
 }